refactor: enum status 통합
