+++ SCM/backend/src/main/java/error/pirate/backend/shippingInstruction/query/dto/ShippingInstructionListRequest.java
@@ -1,5 +1,6 @@
 package error.pirate.backend.shippingInstruction.query.dto;
 
+import error.pirate.backend.shippingInstruction.command.domain.aggregate.entity.ShippingInstructionStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
 import lombok.Data;
