+++ SCM/backend/src/main/java/error/pirate/backend/shippingInstruction/query/mapper/ShippingInstructionMapper.java
@@ -1,5 +1,6 @@
 package error.pirate.backend.shippingInstruction.query.mapper;
 
+import error.pirate.backend.shippingInstruction.command.domain.aggregate.entity.ShippingInstructionStatus;
 import error.pirate.backend.shippingInstruction.query.dto.*;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
